docs/examples/flutter-django/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
streamkit-django>=0.1.0
python-dotenv>=1.0
